+++ EZLogger/HostForms/DueDates1370Host.vb
@@ -1,18 +1,21 @@
-﻿Imports EZLogger.Helpers
-Public Class ReportTypeHost
+﻿Imports System.Windows.Forms.Integration
+Imports Views ' Adjust namespace if needed
 
-	Private Sub ReportTypeHost_Load(sender As Object, e As EventArgs) Handles MyBase.Load
-		Dim view As New ReportTypeView(Me)
+Public Class DueDates1370Host
+
+	Private Sub DueDates1370Host_Load(sender As Object, e As EventArgs) Handles MyBase.Load
+		Dim view As New DueDates1370View()
 		ElementHost1.Child = view
 
 		' Set form size and title
-		Me.ClientSize = New Drawing.Size(450, 375)
+		Me.ClientSize = New Drawing.Size(375, 460)
+		Me.Text = ""
 
 		' Optional UI settings
 		Me.MinimizeBox = False
 		Me.MaximizeBox = False
 		Me.ShowIcon = False
-		'Me.FormBorderStyle = FormBorderStyle.FixedSingle
+		Me.FormBorderStyle = FormBorderStyle.FixedSingle
 		Me.TopMost = True
 
 		' Optional: center the window